+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 123:
 
 * Use unit-test subtree
+* Fix spurious race in websocket close test
 
 --------------------------------------------------------------------------------